Fakeban code
Copy this into the custom command module on dyno.gg!

NOTE: I do not claim to own nor manage the discord bot dyno, the website dyno.gg, or the website discord.com. This is just a command I have created for use within servers where you have permissions to add bots, and I do not take responsibility for your use of this command. You may use this command in your own personal dashboard, as long as you do not claim to have created it yourself.